add-ontology/functions/requirements.txt
firebase-functions
firebase-admin
neo4j
python-dotenv